branch management all apis done - cors issue

+++ clinicX/src/app/Pages/branch/branch.component.ts
@@ -5,14 +5,14 @@ import { TopbarComponent } from '../../Components/topbar/topbar.component';
 import { TableModule } from 'primeng/table';
 import { ButtonModule } from 'primeng/button';
 import { ChartModule } from 'primeng/chart';
-import { FormBuilder, Validators } from '@angular/forms';
+import { FormBuilder, Validators, ReactiveFormsModule } from '@angular/forms';
 import { Router } from '@angular/router';
 import { NgToastService } from 'ng-angular-popup';
 import { ApiService } from '../../Services/api.service';
 import { AuthService } from '../../Services/auth.service';
 import { InputTextModule } from 'primeng/inputtext';
 import { HttpClientModule, HttpClient } from '@angular/common/http';
-import { ReactiveFormsModule } from '@angular/forms';
+import { DialogModule } from 'primeng/dialog';
 
 @Component({
   selector: 'app-branch',
@@ -24,9 +24,16 @@ import { ReactiveFormsModule } from '@angular/forms';
     ButtonModule,
     CommonModule,
     ChartModule,
+    DialogModule,
+    ReactiveFormsModule,
+    HttpClientModule,
+    InputTextModule,
   ],
   templateUrl: './branch.component.html',
-  styleUrl: '../dashboard-home/dashboard-home.component.css',
+  styleUrls: [
+    '../dashboard-home/dashboard-home.component.css',
+    './branch.component.css',
+  ],
 })
 export class BranchComponent {
   constructor(
@@ -38,7 +45,15 @@ export class BranchComponent {
     private apiservice: ApiService
   ) {}
   branches: any;
+  mode: string = 'add';
   selectedbranch: any = [];
+  selectedBranch: any;
+
+  visible: boolean = false;
+
+  showDialog() {
+    this.visible = true;
+  }
 
   ngOnInit() {
     this.fetchbranches();
@@ -62,19 +77,28 @@ export class BranchComponent {
 
   deleteSelectedBranches() {
     // Assuming selectedbranch is an array of branch IDs to be deleted
-    this.apiservice.deleteBranch(this.selectedbranch).subscribe(
-      (response: any) => {
-        // Handle the API response
-        console.log('API response:', response);
 
-        // Refresh the branches after deletion
-        this.fetchbranches();
-      },
-      (error) => {
-        // Handle the API error
-        console.error('API error:', error);
-      }
-    );
+    if (this.selectedbranch.length !== 0) {
+      this.apiservice.deleteBranch(this.selectedbranch).subscribe(
+        (response: any) => {
+          // Handle the API response
+          console.log('API response:', response);
+
+          // Refresh the branches after deletion
+          this.fetchbranches();
+        },
+        (error) => {
+          // Handle the API error
+          console.error('API error:', error);
+        }
+      );
+    } else {
+      this.toast.info({
+        detail: 'Info',
+        summary: 'Select Branch',
+        duration: 5000,
+      });
+    }
   }
 
   class: any = 'p-datatable-sm';
@@ -91,5 +115,167 @@ export class BranchComponent {
       );
     }
     console.log(this.selectedbranch);
+  }
+
+  branchForm = this.fb.group({
+    name: ['', Validators.required],
+    description: ['', Validators.required],
+    address: ['', Validators.required],
+    phone_number: [
+      '',
+      [
+        Validators.required,
+        // Validators.pattern(
+        //   /^(\+254\w?7\d{8}|\+2541\d{8}|\+25420\d{7}|\+25430\d{7}|\+25440\d{7}|\+25450\d{7}|\+25460\d{7}|\+25470\d{7}|\+25480\d{7}|\+25490\d{7})$/
+        // ),
+      ],
+    ],
+    longitude: ['', [Validators.required]],
+    latitude: ['', [Validators.required]],
+  });
+
+  addbranch() {
+    this.mode = 'add';
+
+    if (this.branchForm.valid) {
+      const payload = {
+        id: this.selectedbranch[0],
+        ...this.branchForm.value,
+      };
+
+      console.log(`Payload: ${JSON.stringify(payload)}`);
+
+      this.apiservice.createBranch(payload).subscribe(
+        (response: any) => {
+          // Handle the API response
+          console.log('API response:', response);
+          if (response.success) {
+            this.toast.success({
+              detail: 'Successful',
+              summary: 'Branch created Successful',
+              duration: 2500,
+            });
+
+            setTimeout(() => {
+              this.visible = false;
+            }, 2000);
+            this.route.navigate(['/branchmanagement']);
+          } else {
+            this.toast.error({
+              detail: 'Error',
+              summary: 'Branch Failed. Please try again later.',
+              duration: 5000,
+            });
+
+            setTimeout(() => {
+              this.visible = false;
+            }, 2000);
+          }
+        },
+        (error) => {
+          // Handle the API error
+          console.error('API error:', error);
+          this.toast.error({
+            detail: 'Error',
+            summary: 'Branch Failed. Please try again later.',
+            duration: 5000,
+          });
+          setTimeout(() => {
+            this.visible = false;
+          }, 2000);
+        }
+      );
+    } else {
+      this.toast.error({
+        detail: 'Error',
+        summary: 'Branch Creation Failed. Please check form values.',
+        duration: 5000,
+      });
+    }
+  }
+
+  editBranch(branch: any) {
+    if (this.selectedbranch.length !== 0) {
+      this.visible = true;
+      this.mode = 'edit';
+      console.log(this.mode);
+      if (branch) {
+        this.selectedBranch = { ...branch };
+        this.branchForm.patchValue({
+          name: branch.name,
+          description: branch.description,
+          address: branch.address,
+          phone_number: branch.phone_number,
+          longitude: branch.longitude,
+          latitude: branch.latitude,
+        });
+      } else {
+        this.selectedBranch = null;
+        this.branchForm.reset();
+      }
+    } else {
+      this.toast.info({
+        detail: 'Info',
+        summary: 'Select Branch',
+        duration: 5000,
+      });
+    }
+  }
+
+  updateBranch() {
+    if (this.branchForm.valid) {
+      const payload = {
+        ...this.branchForm.value,
+      };
+
+      console.log(`Payload: ${JSON.stringify(payload)}`);
+
+      this.apiservice.updateBranch(this.selectedbranch, payload).subscribe(
+        (response: any) => {
+          // Handle the API response
+          console.log('API response:', response);
+          if (response.success) {
+            this.toast.success({
+              detail: 'Successful',
+              summary: 'Branch Updated Successful',
+              duration: 2500,
+            });
+
+            setTimeout(() => {
+              this.visible = false;
+            }, 2000);
+            this.route.navigate(['/branchmanagement']);
+          } else {
+            this.toast.error({
+              detail: 'Error',
+              summary: 'Branch Update Failed. Please try again later.',
+              duration: 5000,
+            });
+
+            setTimeout(() => {
+              this.visible = false;
+            }, 2000);
+          }
+        },
+        (error) => {
+          // Handle the API error
+          console.error('API error:', error);
+          this.toast.error({
+            detail: 'Error',
+            summary: 'Branch Failed. Please try again later.',
+            duration: 5000,
+          });
+          setTimeout(() => {
+            this.visible = false;
+          }, 2000);
+        }
+      );
+    } else {
+      this.toast.error({
+        detail: 'Error',
+        summary: 'Branch Failed. Please check form values.',
+        duration: 5000,
+      });
+    }
   }
 }
